+++ src/main/java/example/user/domain/EMail.java
@@ -1,0 +1,8 @@
+package example.user.domain;
+
+@OverarchNode(tags = {"value"})
+public class EMail {
+
+    String email;
+
+}