+++ Changelog.md
@@ -2,6 +2,8 @@
 
 ## Version (NEXT)
 * removed redundant info from annotations
+* added EMail value object
+* added tags to annotations
 
 ## Version 0.1.0
 * initial version